Move Directive into directive.ts
