+++ packages/lit-html/src/directives/until.ts
@@ -12,7 +12,8 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {directive, Part, noChange, _$private} from '../lit-html.js';
+import {Part, noChange, _$private} from '../lit-html.js';
+import {directive} from '../directive.js';
 import {DisconnectableDirective} from '../disconnectable-directive.js';
 
 const DEV_MODE = true;
