+++ packages/lit-html/src/disconnectable-directive.ts
@@ -126,7 +126,6 @@
  */
 
 import {
-  Directive,
   PartInfo,
   NodePart,
   NODE_PART,
@@ -134,7 +133,8 @@ import {
   noChange,
 } from './lit-html.js';
 import {setPartValue} from './directive-helpers.js';
-export {directive} from './lit-html.js';
+export {directive} from './directive.js';
+import {Directive} from './directive.js';
 
 /**
  * Recursively walks down the tree of Parts/TemplateInstances/Directives to set
