+++ packages/lit-html/src/lit-html.ts
@@ -12,6 +12,8 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
+import {Directive} from './directive';
+
 const DEV_MODE = true;
 const ENABLE_EXTRA_SECURITY_HOOKS = true;
 
@@ -324,20 +326,6 @@ export type DirectiveResult<C extends DirectiveClass = DirectiveClass> = {
   values: DirectiveParameters<InstanceType<C>>;
 };
 
-/**
- * Creates a user-facing directive function from a Directive class. This
- * function has the same parameters as the directive's render() method.
- *
- * WARNING: The directive and part API changes are in progress and subject to
- * change in future pre-releases.
- */
-export const directive = <C extends DirectiveClass>(c: C) => (
-  ...values: DirectiveParameters<InstanceType<C>>
-): DirectiveResult<C> => ({
-  _$litDirective$: c,
-  values,
-});
-
 export interface RenderOptions {
   /**
    * An object to use as the `this` value for event listeners. It's often
@@ -404,52 +392,10 @@ let sanitizerFactoryInternal: SanitizerFactory = noopSanitizer;
 
 // Type for classes that have a `_directive` or `_directives[]` field, used by
 // `resolveDirective`
-export type DirectiveParent = AttributePart | NodePart | Directive;
-
-/**
- * Base class for creating custom directives. Users should extend this class,
- * implement `render` and/or `update`, and then pass their subclass to
- * `directive`.
- *
- * WARNING: The directive and part API changes are in progress and subject to
- * change in future pre-releases.
- */
-export abstract class Directive {
-  //@internal
-  _part: NodePart | AttributePart;
-  //@internal
-  _attributeIndex: number | undefined;
-  //@internal
+export interface DirectiveParent {
+  _$parent?: DirectiveParent;
   _directive?: Directive;
-
-  //@internal
-  _$parent: Disconnectable;
-
-  // These will only exist on the DisconnectableDirective subclass
-  //@internal
-  _$disconnetableChildren?: Set<Disconnectable>;
-  //@internal
-  _$setDirectiveConnected?(isConnected: boolean): void;
-
-  constructor(partInfo: PartInfo) {
-    this._$parent = partInfo._$parent;
-    this._part = partInfo._$part;
-    this._attributeIndex = partInfo._$attributeIndex;
-  }
-  /** @internal */
-  _resolve(props: Array<unknown>): unknown {
-    const {_part, _attributeIndex} = this;
-    return resolveDirective(
-      _part,
-      this.update(_part, props),
-      this,
-      _attributeIndex
-    );
-  }
-  abstract render(...props: Array<unknown>): unknown;
-  update(_part: Part, props: Array<unknown>): unknown {
-    return this.render(...props);
-  }
+  _directives?: Array<Directive | undefined>;
 }
 
 /**
