+++ packages/lit-html/src/test/directives/style-map_test.ts
@@ -12,7 +12,8 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {AttributePart, directive, html, render} from '../../lit-html.js';
+import {AttributePart, html, render} from '../../lit-html.js';
+import {directive} from '../../directive.js';
 import {StyleInfo, styleMap} from '../../directives/style-map.js';
 import {assert} from '@esm-bundle/chai';
